Mostly cleanup of File. Added class Path for other file helpers. Some API changes in SelectorListener and IReaderWriter. Also added README files in the test dir that contains the status of the unit tests at this time.

+++ src/Selector.cpp
@@ -97,7 +97,7 @@ void Selector::shutdown()
 	}
 }
 
-void Selector::addListener( int fd, short events, SelectorListener *listener, uint32_t private_data )
+void Selector::addListener( int fd, short events, SelectorListener *listener, unsigned long private_data )
 {
 	TRACE_BEGIN( LOG_LVL_INFO );
